+++ web/src/components/Navbar/Navbar.tsx
@@ -4,7 +4,7 @@ import { selectCartItems } from '@redux/cartSlice'
 import { RxHamburgerMenu, RxCross2 } from 'react-icons/rx'
 import { CiSearch } from 'react-icons/ci'
 import { BsBag } from 'react-icons/bs'
-import { categories } from 'utils/helpers'
+import { categories } from '@lib/categories'
 import Text from '@components/Text'
 import Button from '@components/Button'
 import { NavbarProps } from './Navbar.types'
