+++ web/src/components/ProductCard/ProductCard.module.scss
@@ -4,11 +4,11 @@
   display: flex;
   flex-direction: column;
   cursor: pointer;
-  min-width: $spacing-1000;
+  min-width: $spacing-800;
   height: 100%;
 
   @media screen and (min-width: $breakpoint-xl) {
-    min-width: $spacing-1400;
+    min-width: $spacing-900;
   }
 
   &__image {
