+++ web/src/pages/collection/[slug].tsx
@@ -11,9 +11,9 @@ import ProductsGrid from '@components/ProductsGrid'
 import ProductCard from '@components/ProductCard'
 import Headline from '@components/Headline'
 import SelectMenu from '@components/SelectMenu'
-import styles from './CollectionPage.module.scss'
 import Text from '@components/Text'
 import CategoryAccordion from '@components/CategoryAccordion'
+import styles from './CollectionPage.module.scss'
 
 const CollectionPage = () => {
   const router = useRouter()
@@ -70,14 +70,16 @@ const CollectionPage = () => {
 
   return (
     <div className={styles.Root}>
-      <aside className={styles.Root__sidebar}></aside>
+      <Headline element="h2" size="lg" className={styles.Root__headline}>
+        {slug.replace('-', ' ')}
+      </Headline>
 
-      <main className={styles.Root__main}>
-        <div>
-          <Headline element="h2" size="lg" className={styles.Root__headline}>
-            {slug.replace('-', ' ')}
-          </Headline>
+      <div className={styles.Root__wrapper}>
+        <aside className={styles.Root__sidebar}>
+          <CategoryAccordion />
+        </aside>
 
+        <main className={styles.Root__main}>
           <div className={styles.Root__selectContainer}>
             <Text element="p" size="md" className={styles.Root__selectLabel}>
               Sort by:
@@ -87,15 +89,15 @@ const CollectionPage = () => {
               <SelectMenu options={options} handleOptions={handleOrderBy} />
             </div>
           </div>
-        </div>
 
-        <ProductsGrid>
-          {products &&
-            products.map((product) => (
-              <ProductCard key={product._id} product={product} />
-            ))}
-        </ProductsGrid>
-      </main>
+          <ProductsGrid>
+            {products &&
+              products.map((product) => (
+                <ProductCard key={product._id} product={product} />
+              ))}
+          </ProductsGrid>
+        </main>
+      </div>
     </div>
   )
 }
