+++ web/src/pages/product/[slug].tsx
@@ -1,8 +1,8 @@
 import { GetServerSideProps } from 'next'
 import { useRouter } from 'next/router'
 import { RxChevronLeft } from 'react-icons/rx'
-import { fetchProductById } from 'utils/sanityAPI'
-import { fetchSimilarProducts } from 'utils/sanityAPI'
+import { fetchProductById } from '@utils/sanityAPI'
+import { fetchSimilarProducts } from '@utils/sanityAPI'
 import ProductDetails from '@components/ProductDetails'
 import Button from '@components/Button'
 import Text from '@components/Text'
